fix(UI): Use path URL strategy rather than hash URL strategy
Flutter web application is configured with hash URL strategy by default.
See the following links for details:
. https://docs.flutter.dev/development/ui/navigation/url-strategies
. https://github.com/flutter/flutter/issues/33245

This casued the UI fails in communicating with the backend when it's
behind the reverse proxy as the URL is composed in
```
<ip_address>/#/demo/1.0/games
```
rather than
```
<ip_address>/demo/1.0/games
```

Hence we need to switch using path URL strategy rather than hash URL
strategy.

Also the imported `flutter_web_plugins` depends on `js: ^0.6.3`, we need
to downgrade the js version to make this plugin working for us.