fix(snap): specify the `--base-href` when building the flutter web app
In our case, we would run cloud gaming demo service behind the reverse
proxy and can be accessible with `\demo\` subpath. By default, when
running the `fluter build web` command to build the web app, it always
serves at the root path.
```
<base href="/">
```
To suit our need, we need to configure the base href to `\demo\`
so that the web UI can be loaded properly if we deploy it behind
the reverse proxy.