fix(scripts): fix the broken API download link for bombsquad
As a new version of BombSquad is published, the old one was moved
from the top level folder to the old one of the file sharing website,
this broke the installation of the cloud-gaming demo.

To avoid changing the APK download url frequently whenever a new
release is published, we now use the fixed download URL which points
to 1.17.1 version by default.

Test: run install script